some changes for my personal use

+++ CoinmarketcapClient/CoinmarketcapData.cs
@@ -69,12 +69,12 @@ namespace NoobsMuc.Coinmarketcap.Client
 
     public class CurrenyPriceInfo
     {
-        public double? price { get; set; }
-        public double? volume_24h { get; set; }
-        public double? percent_change_1h { get; set; }
-        public double? percent_change_24h { get; set; }
-        public double? percent_change_7d { get; set; }
-        public double? market_cap { get; set; }
+        public decimal? price { get; set; }
+        public decimal? volume_24h { get; set; }
+        public decimal? percent_change_1h { get; set; }
+        public decimal? percent_change_24h { get; set; }
+        public decimal? percent_change_7d { get; set; }
+        public decimal? market_cap { get; set; }
         public DateTime last_updated { get; set; }
     }
